feat: remove db/rule
