+++ api/script/index.js
@@ -1,32 +1,31 @@
-var bus;
-
+var path = require('path');
 var wrapper = {
     'itemName': function(msg, $meta) {
         $meta.method = 'core.itemName.fetch';
-        return bus.importMethod($meta.method)(msg, $meta);
+        return this.bus.importMethod($meta.method)(msg, $meta);
     },
     'itemCode': function(msg, $meta) {
         $meta.method = 'core.itemCode.fetch';
-        return bus.importMethod($meta.method)(msg, $meta);
+        return this.bus.importMethod($meta.method)(msg, $meta);
     },
     'agentRole': function(msg, $meta) {
         $meta.method = 'db/integration.agentRole.fetch';
-        return bus.importMethod($meta.method)(msg, $meta);
+        return this.bus.importMethod($meta.method)(msg, $meta);
     },
     'accountAlias': function(msg, $meta) {
         $meta.method = 'db/integration.alias.list';
-        return bus.importMethod($meta.method)(msg, $meta);
+        return this.bus.importMethod($meta.method)(msg, $meta);
     },
     'organization': function(msg, $meta) {
-        $meta.method = 'customer.organization.fetch';
-        return bus.importMethod($meta.method)(msg, $meta).then(result => {
+        var $newMeta = Object.assign($meta, { method: 'customer.organization.fetch' });
+        return this.bus.importMethod($meta.method)(msg, $newMeta).then(result => {
             let organization = result.organization;
             return {items: organization.map(v => ({ type: 'organization', value: v.actorId, display: v.organizationName }))};
         });
     },
     'role': function(msg, $meta) {
         $meta.method = 'user.role.fetch';
-        return bus.importMethod($meta.method)(msg, $meta).then(result => {
+        return this.bus.importMethod($meta.method)(msg, $meta).then(result => {
             let role = result.role;
             return {items: role.map(v => ({ type: 'role', value: v.actorId, display: v.name }))};
         });
@@ -34,15 +33,36 @@ var wrapper = {
 };
 
 module.exports = {
+    schema: [
+        {path: path.join(__dirname, '../sql/schema'), linkSP: true},
+        {path: path.join(__dirname, '../sql/schema/seeds')}
+    ],
+    'decision.lookup.response.receive': result => {
+        if (result && Array.isArray(result.split)) {
+            result.split.forEach(split => {
+                if (split.analytics && Array.isArray(split.analytics.rows)) {
+                    split.analytics = split.analytics.rows.reduce((prev, cur) => {
+                        prev[cur.name] = cur.value;
+                        return prev;
+                    }, {});
+                } else if (split.analytics && split.analytics.rows && split.analytics.rows.name && split.analytics.rows.value) {
+                    var analytics = {};
+                    analytics[split.analytics.rows.name] = split.analytics.rows.value;
+                    split.analytics = analytics;
+                }
+                return split;
+            });
+        }
+        return result;
+    },
     'item.fetch': function(msg, $meta) {
-        bus = this.bus;
         var pending = [];
 
         Object.keys(msg).forEach(function(method) {
             if (wrapper[method] !== undefined && msg[method] && msg[method].length > 0) {
-                pending.push(wrapper[method]({alias: msg[method]}, Object.assign({}, $meta)));
+                pending.push(wrapper[method].call(this, {alias: msg[method]}, Object.assign({}, $meta)));
             }
-        });
+        }, this);
 
         return Promise.all(pending).then(function(result) {
             var data = [];
