+++ ui/react/pages/Detail/actions.js
@@ -2,7 +2,7 @@ import * as actionTypes from './actionTypes';
 
 export const fetchRules = (params) => ({
     type: actionTypes.FETCH_RULES,
-    method: 'db/rule.rule.fetch',
+    method: 'rule.rule.fetch',
     params: params || {}
 });
 
@@ -14,7 +14,7 @@ export const fetchNomenclatures = (params) => ({
 
 export const createRule = (params) => ({
     type: actionTypes.CREATE_RULE,
-    method: 'db/rule.rule.add',
+    method: 'rule.rule.add',
     params: params || {}
 });
