+++ ui/react/pages/Main/actionCreators.js
@@ -13,7 +13,7 @@ export const updatePagination = (params) => ({type: actionTypes.updatePagination
 export function fetchRules(params) {
     return {
         type: actionTypes.fetchRules,
-        method: 'db/rule.rule.fetch',
+        method: 'rule.rule.fetch',
         params: params || {}
     };
 };
@@ -22,7 +22,7 @@ export function removeRules(params) {
     return function(dispatch) {
         return dispatch({
             type: actionTypes.removeRules,
-            method: 'db/rule.rule.remove',
+            method: 'rule.rule.remove',
             params: params || {}
         }).then((result) => {
             if (result.error) {
@@ -202,7 +202,7 @@ export function editRule(params) {
 
         return dispatch({
             type: actionTypes.addRule,
-            method: 'db/rule.rule.edit',
+            method: 'rule.rule.edit',
             params: modifiedParams || {}
         }).then((result) => {
             if (result.error) {
@@ -360,7 +360,7 @@ export function addRule(params) {
 
         return dispatch({
             type: actionTypes.addRule,
-            method: 'db/rule.rule.add',
+            method: 'rule.rule.add',
             params: modifiedParams || {}
         }).then((result) => {
             if (result.error) {
